+++ src/lib/helpers/hierarchy.svelte.js
@@ -1,3 +1,4 @@
+import { hierarchyLogger } from '../logger.ts'
 /**
  * @module helpers/hierarchy
  *